Move volume commands to its own class
