+++ src/main/java/de/maxhenkel/audioplayer/AudioPlayer.java
@@ -43,6 +43,7 @@ public class AudioPlayer implements ModInitializer {
                     UploadCommands.class,
                     ApplyCommands.class,
                     UtilityCommands.class,
+                    VolumeCommands.class,
                     PlayCommands.class
             ).setPermissionManager(AudioPlayerPermissionManager.INSTANCE).build();
         });
