+++ src/main/java/de/maxhenkel/audioplayer/command/UtilityCommands.java
@@ -16,9 +16,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.item.*;
 
-import javax.annotation.Nullable;
 import java.util.Optional;
-import java.util.UUID;
 
 @Command("audioplayer")
 public class UtilityCommands {
@@ -64,47 +62,6 @@ public class UtilityCommands {
         context.getSource().sendSuccess(() -> Component.literal("Successfully cleared item"), false);
     }
 
-    @RequiresPermission("audioplayer.volume")
-    @Command("volume")
-    public void volumeWithId(CommandContext<CommandSourceStack> context, @Name("id") UUID uuid, @OptionalArgument @Name("volume") @Min("0.00") @Max("100.00") Float volume) throws CommandSyntaxException {
-        volumeCommand(context,uuid,volume);
-    }
-
-    @RequiresPermission("audioplayer.volume")
-    @Command("volume")
-    public void volumeHeldItem(CommandContext<CommandSourceStack> context, @OptionalArgument @Name("volume") @Min("0.00") @Max("100.00") Float volume) throws CommandSyntaxException {
-        CustomSound customSound = getHeldSound(context);
-        if (customSound == null) {
-            return;
-        }
-        volumeCommand(context,customSound.getSoundId(),volume);
-    }
-
-    private void volumeCommand(CommandContext<CommandSourceStack> context, UUID id, @Nullable Float volume) {
-        if (!AudioManager.checkSoundExists(context.getSource().getServer(),id)) {
-            context.getSource().sendFailure(Component.literal("Sound does not exist"));
-            return;
-        }
-        Optional<VolumeOverrideManager> optionalMgr = VolumeOverrideManager.instance();
-        if (optionalMgr.isEmpty()) {
-            context.getSource().sendFailure(Component.literal("An internal error occurred"));
-            return;
-        }
-        VolumeOverrideManager mgr = optionalMgr.get();
-        if (volume == null) {
-            float currentVolume = mgr.getAudioVolume(id);
-            context.getSource().sendSuccess(() -> Component.literal("Current volume is %.2f".formatted(currentVolume * 100f)), false);
-        } else {
-            if (volume == 100.00f) {
-                // will remove volume from json, to keep json file smaller
-                mgr.setAudioVolume(id,null);
-            }
-            mgr.setAudioVolume(id,volume / 100f);
-            AudioPlayer.AUDIO_CACHE.remove(id);
-            context.getSource().sendSuccess(() -> Component.literal("Set volume, this will apply next time the sound plays"), false);
-        }
-    }
-
     @Command("id")
     public void id(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
         CustomSound customSound = getHeldSound(context);
@@ -142,7 +99,7 @@ public class UtilityCommands {
         })), false);
     }
 
-    private static CustomSound getHeldSound(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
+    public static CustomSound getHeldSound(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
         ServerPlayer player = context.getSource().getPlayerOrException();
         ItemStack itemInHand = player.getItemInHand(InteractionHand.MAIN_HAND);
